Rename log_grid to information_grid
