+++ src/components/details_view_components/information_grid.rs
